==> Final Code Notebooks/03_Joining_Data.ipynb <==
In this notebook different files from previous notebooks and ones downloaded separately are loaded, preprocessed and joined so that final analysis could be done on them. 

import pandas as pd 
import numpy as np 
import re
from pandasql import sqldf 

df_articles: contains the information for each article extracted from wikimedia

df_ores: contains the prediction for each article and its rev_id 

df_population: contains the population estimates for 2022 for each state 

df_states: contains the region wise division of each state

df_articles = pd.read_csv('page_record.csv')
df_ores = pd.read_csv('ores_predictions.csv')
df_population = pd.read_excel('NST-EST2022-POP.xlsx')
df_states = pd.read_excel('US States by Region - US Census Bureau.xlsx')

Preprocessing the population record dataframe. Primarily to correct the columns name issue as can be seen below and secondly there was a . before some names of states. So I removed those as we will be joining on these in future. 

df_population.head()

# Set the first row as column names
df_population.columns = df_population.iloc[0]

# Drop the first row (which is now used as column names)
df_population = df_population[1:]

# Reset the index (optional)
df_population.reset_index(drop=True, inplace=True)

# Display the DataFrame
df_population.tail()

# Remove dots from the values in column 'A'
df_population['Geographic Area'] = df_population['Geographic Area'].str.replace('.', '', regex=False)

# Display the modified DataFrame
df_population.head()

df_population[df_population['Geographic Area'] == 'Alabama'][0:8]

Preprocessing the articles dataframe to separate the cities and states. A regex expression is written to ensure that state has an exact state and any entries before the state are just entered into the city column

df_articles.head()

# Extract city and state from the 'Location' column
df_articles['City'] = df_articles['title'].str.split(',').str[0]
df_articles['State'] = df_articles['title'].str.split(',').str[-1]

df_articles[['City', 'State']] = df_articles[['City', 'State']].applymap(lambda x: x.strip())

#checking that how many unique values are in the states column
df_articles['State'].unique()

df_articles and df_ores are joined on lastrevid and rev_id respectively and then a few columns are dropped. Rest all irrelevant columns will be dropped towards the end before saving the file.

join_1 = pd.merge(df_articles, df_ores, left_on=df_articles['lastrevid'],
                  right_on = df_ores['rev_id'], how='left')

join_1.columns

join_1.drop(['key_0', 'Unnamed: 0_x'], axis = 1, inplace = True)

join_1[0:20]

Over here our files from the above join are joined with df_population which has data of state wise population for 2022. A few columns are dropped

join_2 = pd.merge(join_1, df_population, left_on = join_1['State'],  right_on = df_population['Geographic Area'], 
                  how='left')

join_2.columns

join_2.drop(['April 1, 2020 Estimates Base', 2020, 2021, 'key_0' , 'Unnamed: 0_y'],
            axis = 1, inplace = True)

join_2.head()

join_2.columns

join_2.to_csv('join_with_states.csv')

df_states.head()

This is the final join with df_states to get the region and division which we will be using in our final anaylsis

join_3 = pd.merge(join_2, df_states, left_on = join_2['State'],  right_on = df_states['STATE'], 
                  how='left')


join_3.head()

join_3.columns

join_3.columns

ALL IRRELEVANT COLUMNS ARE NOW DROPPED TO BRING OUR TABLE TO DESIRED FORMAT. ANY MISSING VALUES ARE ALSO DROPPED 

join_3.drop(['pageid', 'ns', 'touched', 'length', 'talkid', 'fullurl', 'editurl', 'canonicalurl', 
             'watchers', 'redirect', 'new', 'City', 'lastrevid', 'Geographic Area', 'STATE', 'REGION', 'key_0',
            'contentmodel', 'pagelanguage', 'pagelanguagehtmlcode', 'pagelanguagedir'], 
            axis = 1, inplace = True)

join_3.head()

final_df = join_3.dropna()

##Rows with missing values 
print(f'rows with missing values are: {join_3.shape[0] - final_df.shape[0]}')

final_df = final_df.rename(columns={2022: 'population'})

final_df.to_csv('wp_scored_city_articles_by_state.csv')
