==> Final Code Notebooks/04_Results.ipynb <==
# Final jupyter notebook to display the resulting tables as required

import pandas as pd 
import numpy as np
import matplotlib.pyplot as plt 

df = pd.read_csv('wp_scored_city_articles_by_state.csv')

df.drop('Unnamed: 0', axis = 1 , inplace = True)

df.head()

# RESULT 1

# Group the DataFrame by 'State' and count the number of titles for each state
state_title_counts = df.groupby('State')['title'].count().reset_index()

# Rename the columns for clarity
state_title_counts.columns = ['State', 'Total Titles']

# Display the result
state_title_counts.head()

unique_pop_state_pairs = df[['population', 'State']].drop_duplicates()

df_merged = pd.merge(state_title_counts, unique_pop_state_pairs, left_on = state_title_counts['State'],  
                  right_on = unique_pop_state_pairs['State'], how='left')
df_merged.drop(['key_0', 'State_y'], axis =1 , inplace = True)
df_merged['articles per capita'] = df_merged['Total Titles'] / df_merged['population']
df_merged = df_merged.sort_values(by='articles per capita' , ascending = False).reset_index(drop = True)
print('Top 10 US states by coverage: The 10 US states with the highest total articles per capita (in descending order)')
df_merged[0:10]

# RESULT 2

print('Bottom 10 US states by coverage: The 10 US states with the lowest total articles per capita (in ascending order) .')
df_merged[24:35].sort_values(by = 'articles per capita').reset_index(drop = True)

# RESULT 3

# Define a function to map Prediction values to quality
def map_quality(prediction):
    if prediction in ['FA', 'GA']:
        return 'High Quality'
    else:
        return 'Average'

# Apply the map_quality function to create a new 'Quality' column
df['Quality'] = df['Prediction'].apply(lambda x: map_quality(x))

df.head()

# Group the DataFrame by 'State' and count high quality and average titles
state_quality_counts = df.groupby('State')['Quality'].value_counts().unstack(fill_value=0)

# Display the result
state_quality_counts.head()

# Reset the index into a regular column
state_quality_counts.reset_index(drop=False, inplace=True)

#state_quality_counts.drop('level_0', axis = 1, inplace = True)
state_quality_counts.columns

# Display the DataFrame with the index as a column
state_quality_counts.head()

unique_pop_state_pairs = df[['population', 'State']].drop_duplicates()

df_merged_2 = pd.merge(state_quality_counts, unique_pop_state_pairs, left_on = state_quality_counts['State'],  
                  right_on = unique_pop_state_pairs['State'], how='left')
df_merged_2.drop(['key_0', 'State_y'], axis = 1, inplace = True)

df_merged_2.head()

df_merged_2['High Quality articles per capita'] = df_merged_2['High Quality'] / df_merged_2['population']
df_merged_2 = df_merged_2.sort_values(by='High Quality articles per capita' , ascending = False).reset_index(drop = True)

print('Top 10 US states by high quality: The 10 US states with the highest high quality articles per capita (in descending order) .')
df_merged_2[0:10]

# RESULT 4

print('Bottom 10 US states by high quality: The 10 US states with the lowest high quality articles per capita (in ascending order).')

df_merged_2[24:35].sort_values(by = 'High Quality articles per capita').reset_index(drop = True)



# RESULT 5

# Group the DataFrame by 'DIVISION' and count the number of titles for each state
division_title_counts = df.groupby('DIVISION')['title'].count().reset_index()

# Rename the columns for clarity
division_title_counts.columns = ['DIVISION', 'Total Titles']

# Display the result
division_title_counts.head()

unique_pop_state_div_pairs = df[['population', 'State', 'DIVISION']].drop_duplicates()
unique_pop_state_div_pairs.head()

division_population = df.groupby('DIVISION')['population'].sum().reset_index()
division_population.head()

df_merged_3 = pd.merge(division_title_counts, division_population, left_on = division_title_counts['DIVISION'],  
                  right_on = division_population['DIVISION'], how='left')
df_merged_3.drop(['key_0', 'DIVISION_y'], axis = 1, inplace = True)


df_merged_3

df_merged_3['articles per capita'] = df_merged_3['Total Titles'] / df_merged_3['population']
df_merged_3 = df_merged_3.sort_values(by='articles per capita' , ascending = False).reset_index(drop = True)
print('Census divisions by total coverage: A rank ordered list of US census divisions (in descending order) by total articles per capita.')
df_merged_3

# RESULT 6

# Group the DataFrame by 'State' and count high quality and average titles
division_quality_counts = df.groupby('DIVISION')['Quality'].value_counts().unstack(fill_value=0)

# Display the result
division_quality_counts.head()

# Reset the index into a regular column
division_quality_counts.reset_index(drop=False, inplace=True)

#state_quality_counts.drop('level_0', axis = 1, inplace = True)
division_quality_counts.columns

division_quality_counts.head()

df_merged_4 = pd.merge(division_quality_counts, division_population, left_on = division_quality_counts['DIVISION'],  
                  right_on = division_population['DIVISION'], how='left')

df_merged_4['High Quality articles per capita'] = df_merged_4['High Quality'] / df_merged_4['population']
df_merged_4 = df_merged_4.sort_values(by='High Quality articles per capita' , ascending = False).reset_index(drop = True)
df_merged_4.drop(['key_0', 'DIVISION_y'], axis = 1, inplace = True)
print('Census divisions by high quality coverage: Rank ordered list of US census divisions (in descending order) by high quality articles per capita.')
df_merged_4
